Combat and ai updates (#141)
* Added Medium Enemy AI control. Made combat updates.

* Implemented Heavy and Light aiming.

* Added snap aiming and heavy AI behavior

* Enemies move forward now and medium rolling away works

* We did it squad fam

* ree meta

* Removed Artifact

+++ Space Core/Assets/Resources/Scripts/Environment/DestructibleWall.cs
@@ -13,7 +13,7 @@ public class DestructibleWall : MonoBehaviour
         {
             if (other.gameObject.GetComponent<Bullet>().Source == DamageSource.Player)
             {
-                TakeDamage(other.gameObject.GetComponent<Bullet>().Damage);
+                TakeDamage((int) other.gameObject.GetComponent<Bullet>().Damage);
                 other.gameObject.SetActive(false);
             }
         }
